requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
art==6.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
editor==1.6.6
exceptiongroup==1.2.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
inquirer==3.3.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.11
langchain-core==0.2.23
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langsmith==0.1.93
lxml==5.2.2
multidict==6.0.5
numpy==1.26.4
openai==1.37.0
orjson==3.10.6
packaging==24.1
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pypng==0.20220715.0
pytest==8.3.2
python-dotenv==1.0.1
PyYAML==6.0.1
readchar==4.1.0
regex==2024.5.15
requests==2.32.3
runs==1.2.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
tenacity==8.5.0
tiktoken==0.7.0
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
xmod==1.8.1
yarl==1.9.4
